Clean up ServiceTemplatesDialog
