+++ services/orchest-webserver/client/src/components/ServiceTemplatesDialog/index.js
@@ -1,0 +1,2 @@
+// @ts-check
+export * from "./ServiceTemplatesDialog";
